+++ space_invaders/textures/bullets_texture.h
@@ -3,7 +3,7 @@
 
 #include "stdint.h"
 
-static uint16_t bullets_bits[] = {
+static uint32_t bullets_bits[] = {
 /* MY_BULLET   ():
    ht=4, width=1
    +-+
@@ -12,10 +12,10 @@ static uint16_t bullets_bits[] = {
    |*|
    |*|
    +-+ */
-0x1,
-0x1,
-0x1,
-0x1,
+        0x1,
+        0x1,
+        0x1,
+        0x1,
 
 /* BULLET_1 cond_1  :
    ht=7, width=3
@@ -28,13 +28,13 @@ static uint16_t bullets_bits[] = {
    |  *|
    | * |
    +---+ */
-0x2,
-0x1,
-0x2,
-0x4,
-0x2,
-0x1,
-0x2,
+        0x2,
+        0x1,
+        0x2,
+        0x4,
+        0x2,
+        0x1,
+        0x2,
 
 /* BULLET_1 cond_2  :
    ht=7, width=3
@@ -47,13 +47,13 @@ static uint16_t bullets_bits[] = {
    | * |
    |*  |
    +---+ */
-0x1,
-0x2,
-0x4,
-0x2,
-0x1,
-0x2,
-0x4,
+        0x1,
+        0x2,
+        0x4,
+        0x2,
+        0x1,
+        0x2,
+        0x4,
 
 /* BULLET_1 cond_3  :
    ht=7, width=3
@@ -66,13 +66,13 @@ static uint16_t bullets_bits[] = {
    |*  |
    | * |
    +---+ */
-0x2,
-0x4,
-0x2,
-0x1,
-0x2,
-0x4,
-0x2,
+        0x2,
+        0x4,
+        0x2,
+        0x1,
+        0x2,
+        0x4,
+        0x2,
 
 
 
@@ -87,13 +87,13 @@ static uint16_t bullets_bits[] = {
    | * |
    |  *|
    +---+ */
-0x4,
-0x2,
-0x1,
-0x2,
-0x4,
-0x2,
-0x1,
+        0x4,
+        0x2,
+        0x1,
+        0x2,
+        0x4,
+        0x2,
+        0x1,
 
 /* BULLET_2 cond_1  :
    ht=7, width=3
@@ -106,13 +106,13 @@ static uint16_t bullets_bits[] = {
    | * |
    | * |
    +---+ */
-0x2,
-0x2,
-0x2,
-0x2,
-0x7,
-0x2,
-0x2,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
+        0x7,
+        0x2,
+        0x2,
 
 /* BULLET_2 cond_2  :
    ht=7, width=3
@@ -125,13 +125,13 @@ static uint16_t bullets_bits[] = {
    | * |
    | * |
    +---+ */
-0x2,
-0x2,
-0x7,
-0x2,
-0x2,
-0x2,
-0x2,
+        0x2,
+        0x2,
+        0x7,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
 
 /* BULLET_2 cond_3  :
    ht=7, width=3
@@ -144,13 +144,13 @@ static uint16_t bullets_bits[] = {
    | * |
    | * |
    +---+ */
-0x7,
-0x2,
-0x2,
-0x2,
-0x2,
-0x2,
-0x2,
+        0x7,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
 
 /* BULLET_2 cond_4  :
    ht=7, width=3
@@ -163,13 +163,13 @@ static uint16_t bullets_bits[] = {
    | * |
    |***|
    +---+ */
-0x2,
-0x2,
-0x2,
-0x2,
-0x2,
-0x2,
-0x7,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
+        0x2,
+        0x7,
 
 /* BULLET_3 cond_1  :
    ht=7, width=3
@@ -182,13 +182,13 @@ static uint16_t bullets_bits[] = {
    |** |
    | **|
    +---+ */
-0x2,
-0x2,
-0x6,
-0x3,
-0x2,
-0x6,
-0x3,
+        0x2,
+        0x2,
+        0x6,
+        0x3,
+        0x2,
+        0x6,
+        0x3,
 
 /* BULLET_3 cond_2  :
    ht=7, width=2
@@ -201,13 +201,13 @@ static uint16_t bullets_bits[] = {
    | *|
    | *|
    +--+ */
-0x1,
-0x1,
-0x1,
-0x1,
-0x1,
-0x1,
-0x1,
+        0x1,
+        0x1,
+        0x1,
+        0x1,
+        0x1,
+        0x1,
+        0x1,
 
 /* BULLET_3 cond_3  :
    ht=7, width=3
@@ -220,13 +220,14 @@ static uint16_t bullets_bits[] = {
    | * |
    | * |
    +---+ */
-0x3,
-0x6,
-0x2,
-0x3,
-0x6,
-0x2,
-0x2,
+        0x3,
+        0x6,
+        0x2,
+        0x3,
+        0x6,
+        0x2,
+        0x2,
+
 };
 
 #endif //SPACE_INVADERS_BULLETS_TEXTURE_H
